requirements.txt
slack-sdk==3.36.0
anthropic==0.18.1
python-dotenv==1.0.0
asyncio-mqtt==0.16.1
websockets==12.0
aiohttp>=3.9.1
pydantic==2.5.0
loguru==0.7.2
colorama==0.4.6
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
numpy==1.24.3 
httpx<=0.27